# BookShop/src/main/java/com/jaz/bookshop/dto/Author.java
package com.jaz.bookshop.dto;

import jakarta.persistence.*;
import lombok.Data;

@Data
@Entity
@Table(name = "authors")
public class Author {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;
    private String name;

}